@@ -2,20 +2,14 @@ package org.bobstuff.bobboardview.app.scrum
 
 import android.graphics.Color
 import android.os.Bundle
-import android.support.v7.app.AppCompatActivity
-import android.support.v7.widget.RecyclerView
-import android.support.v7.widget.Toolbar
-import android.util.Log
+import androidx.appcompat.app.AppCompatActivity
+import androidx.recyclerview.widget.RecyclerView
+import androidx.appcompat.widget.Toolbar
 import android.util.TypedValue
 import android.view.MenuItem
 import android.view.View
-import android.widget.Toast
 import org.bobstuff.bobboardview.*
-import org.bobstuff.bobboardview.BobBoardSimpleDividersOrientation.HORIZONTAL
 import org.bobstuff.bobboardview.app.R
-import org.bobstuff.bobboardview.app.simple.SimpleListAdapter
-import org.bobstuff.bobboardview.app.trello.model.BoardList
-import org.bobstuff.bobboardview.app.util.DragOperation
 import org.bobstuff.bobboardview.app.util.getWindowSize
 import org.bobstuff.bobboardview.app.util.setStatusBarColor
 import org.bobstuff.bobboardview.app.util.styleToolbarContent
