@@ -1,8 +1,8 @@
 package org.bobstuff.bobboardview.app.simple
 
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import android.support.v7.widget.RecyclerView
+import androidx.recyclerview.widget.RecyclerView
 import android.view.View
 import org.bobstuff.bobboardview.*
 import org.bobstuff.bobboardview.app.R
