@@ -1,6 +1,6 @@
 package org.bobstuff.bobboardview
 
-import android.support.v7.widget.RecyclerView
+import androidx.recyclerview.widget.RecyclerView
 import android.util.Log
 import android.view.View
 
