@@ -1,7 +1,7 @@
 package org.bobstuff.bobboardview
 
 import android.graphics.Rect
-import android.support.v7.widget.RecyclerView
+import androidx.recyclerview.widget.RecyclerView
 import android.view.View
 
 /**
@@ -19,9 +19,9 @@ class BobBoardSimpleDividers(private val dividerSizeInPixels: Int,
                              private val orientation: BobBoardSimpleDividersOrientation):
         RecyclerView.ItemDecoration() {
 
-    override fun getItemOffsets(outRect: Rect, view: View, parent: RecyclerView, state: RecyclerView.State?) {
+    override fun getItemOffsets(outRect: Rect, view: View, parent: RecyclerView, state: RecyclerView.State) {
         val position = parent.getChildAdapterPosition(view)
-        val itemCount = parent.adapter.itemCount
+        val itemCount = parent.adapter!!.itemCount
 
         when (orientation) {
             BobBoardSimpleDividersOrientation.HORIZONTAL -> getHorizontalItemOffsets(outRect, position, itemCount)
