@@ -1,12 +1,8 @@
 package org.bobstuff.bobboardview
 
-import android.graphics.Rect
-import android.support.v7.widget.LinearLayoutManager
-import android.support.v7.widget.RecyclerView
+import androidx.recyclerview.widget.LinearLayoutManager
+import androidx.recyclerview.widget.RecyclerView
 import android.util.Log
-import android.view.View
-import android.widget.Toast
-import java.util.*
 
 /**
  * Created by bob
